my-protfilo/requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.3.0
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.116.2
greenlet==3.2.4
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.3
passlib==1.7.4
psycopg2==2.9.10
pyasn1==0.6.1
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.36.0

# ================= CORE API =================
fastapi>=0.110,<1.0
uvicorn[standard]>=0.25,<1.0
starlette>=0.36,<1.0
pydantic>=2.7,<3.0
pydantic-settings>=2.0,<3.0
python-multipart>=0.0.9
orjson>=3.9

# ================= DATA =================
numpy>=1.26,<2.0
pandas>=2.1
scipy>=1.11
scikit-learn>=1.4

# ================= NLP / RAG =================
langchain>=0.2
langchain-community>=0.2
langchain-core>=0.2
langchain-openai>=0.1
langchain-pinecone>=0.1
langsmith>=0.1
tiktoken>=0.5
sentencepiece>=0.2
langchain-nvidia-ai-endpoints>=0.1
transformers>=4.40
huggingface-hub>=0.22
aiofiles
# ================= VECTOR DB =================
pinecone>=3.0

# ================= CV =================
opencv-python>=4.9
ultralytics>=8.2
cvzone>=1.6

# ================= TORCH =================
torch>=2.1
torchvision>=0.16
torchaudio>=2.1


SQLAlchemy>=2.0
alembic>=1.12
psycopg2-binary>=2.9

# ================= AUTH =================
python-jose>=3.3
passlib[bcrypt]>=1.7
bcrypt>=4.0
email-validator>=2.0

# ================= UTIL =================
python-dotenv>=1.0
requests>=2.31
httpx>=0.27
rich>=13.7
tqdm>=4.66

